+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
-## [2-beta41 - unreleased]
+## [2-beta41 - 2025-04-28]
 - Use <detail> to render stack traces compactly in the HTML view (PR #245) - by @adityaathalye
+- updated Calva custom REPL commands:
+  - shortented the command names
+  - added a few commands from the `snippets` API
+  - switched to using fully qualified `clojure.core/require` 
+- bugfix: adapted the (experimental) `make-hiccup` API to recent format changes
+- bugfix: fixed a typo in the `snippets/make-form-quarto-html!` function
 
 ## [2-beta40 - 2025-04-18]
 - calva custom REPL commands - added "make top level form"